% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_last_update.R
\name{check_aws_last_update}
\alias{check_aws_last_update}
\title{Process LSI-ELOG data.}
\usage{
check_aws_last_update(dirAWS)
}
\arguments{
\item{dirAWS}{full path to the directory to store the parsed data.
Example: "E:/MeteoRwanda_v2/AWS_DATA"}
}
\value{
a data.frame
}
\description{
Get the data from Gidas database, parse and convert to a list object then upload it to "data-int" server.
}
