% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{download.rema}
\alias{download.rema}
\title{Download REMA data from the FTP server.}
\usage{
download.rema(dirAWS)
}
\arguments{
\item{dirAWS}{full path to the directory to store the parsed data.
Example: "E:/MeteoRwanda"}
}
\description{
Download REMA data from the FTP server.
}
\details{
The downloaded data are stored under "E:/MeteoRwanda/AWS_DATA/REMA/ELAB" if \code{dirAWS} is "E:/MeteoRwanda"
}
