% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.lsi.xlog}
\alias{process.lsi.xlog}
\title{Process LSI-XLOG data.}
\usage{
process.lsi.xlog(dirLSIXLOG, dirAWS, dirUP = NULL, upload = TRUE)
}
\arguments{
\item{dirLSIXLOG}{full path to the directory of X-LOG.
Example: "C:/DATA/X-LOG"}

\item{dirAWS}{full path to the directory to store the parsed data.
Example: "E:/MeteoRwanda"}

\item{dirUP}{full path to the directory to store the uploaded data in \code{data-int}.
Default NULL, must be provided if \code{upload} is \code{TRUE}.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA/RAW"}

\item{upload}{logical, if TRUE the data will be uploaded to \code{data-int}}
}
\description{
Read the data from X-LOG storage, parse and convert to a list object then upload it to "data-int" server.
}
