% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.rema_arch}
\alias{process.rema_arch}
\title{Process REMA data archive mode.}
\usage{
process.rema_arch(
  start_min,
  end_min,
  dirELAB,
  dirAWS,
  dirUP = NULL,
  upload = TRUE
)
}
\arguments{
\item{start_min}{the start time to process in the format "YYYY-MM-DD HH:MM".
Example: "2019-12-15 12:50"}

\item{end_min}{the end time to process in the format "YYYY-MM-DD HH:MM"}

\item{dirELAB}{full path to the directory of the downloaded REMA elaborated data from FTP.
Example: "E:/MeteoRwanda/AWS_DATA/REMA/ELAB"}

\item{dirAWS}{full path to the directory to store the parsed data.
Example: "E:/MeteoRwanda"}

\item{dirUP}{full path to the directory to store the uploaded data in \code{data-int}.
Default NULL, must be provided if \code{upload} is \code{TRUE}.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA/RAW"}

\item{upload}{logical, if TRUE the data will be uploaded to \code{data-int}}
}
\description{
Get the REMA data, parse and convert to a list object then upload it to "data-int" server.
}
