% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{upload.logFiles}
\alias{upload.logFiles}
\title{Upload log files to "data-int" server.}
\usage{
upload.logFiles(dirAWS, dirUP)
}
\arguments{
\item{dirAWS}{full path to the directory to store the parsed data.
Example: "E:/MeteoRwanda"}

\item{dirUP}{full path to the directory to store the uploaded data in \code{data-int}.}
}
\description{
Upload log files to "data-int" server.
}
